Add common/version-util.js
